Add Pulls and Issues APIs
